Miscellaneous code cleanups and fixes: - Removed references to unused include directories from Geometry/MdfModel projects. - Updated Foundation/MapGuideCommon/PlatformBase projects to use CPP include files instead of precompiled headers. - Got rid of some compiling warnings in Stylization project.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1071 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Stylization/complex_polygon_gd.cpp
@@ -102,7 +102,7 @@ void complex_polygon_gd::FillPolygon(Point * VertexList, int nVerts, int* Contou
 
     /* Get enough memory to store the entire edge table */
     /* TIS -- reuse a buffer owned by the class */
-    if (m_nBufEdgeState < nVerts)
+    if (m_nBufEdgeState < (size_t)nVerts)
     {
         delete [] m_bufEdgeState;
         m_nBufEdgeState = nVerts;
